Prevent unpacker auto from running when there is no archive

+++ steps/unpacker.rb
@@ -12,34 +12,36 @@ class Unpacker < LBT::StepsFabricator
 
 		def run
 
-			Dir.chdir $global_state.project_dir
+			if not @library.archive.nil?  and not File.exist? @library.archive
+				Dir.chdir $global_state.project_dir
 
-			FileUtils.mkdir_p $global_state.build_dir
+				FileUtils.mkdir_p $global_state.build_dir
 
 
-			@library.work_dir = "#{$global_state.build_dir}/#{@library.name}"
-			FileUtils.rm_rf @library.work_dir if Dir.exist? @library.work_dir
-			FileUtils.mkdir_p @library.work_dir
+				@library.work_dir = "#{$global_state.build_dir}/#{@library.name}"
+				FileUtils.rm_rf @library.work_dir if Dir.exist? @library.work_dir
+				FileUtils.mkdir_p @library.work_dir
 
-			scriptSuccess = false
-			if(@library.archive.include? '.tar' or @library.archive.include? 'tgz')
-				scriptSuccess = untar
-			elsif(@library.archive.include? '.zip')
-				scriptSuccess = unzip
-			end
-
-			throw "Unpack script failed for #{@library.name}" if not scriptSuccess
-
-			# TODO : Verify if there's not a better place to discover if a subdir is needed.
-			if @library.build_subdir.nil?
-				to_ignore = ["__MACOSX", ".", ".."]
-				listing = Dir.entries @library.work_dir
-				to_ignore.each do |ignored|
-					listing.delete ignored
+				scriptSuccess = false
+				if(@library.archive.include? '.tar' or @library.archive.include? 'tgz')
+					scriptSuccess = untar
+				elsif(@library.archive.include? '.zip')
+					scriptSuccess = unzip
 				end
-				@library.build_subdir = (listing.first) if listing.length == 1
-			end
 
+				throw "Unpack script failed for #{@library.name}" if not scriptSuccess
+
+				# TODO : Verify if there's not a better place to discover if a subdir is needed.
+				if @library.build_subdir.nil?
+					to_ignore = ["__MACOSX", ".", ".."]
+					listing = Dir.entries @library.work_dir
+					to_ignore.each do |ignored|
+						listing.delete ignored
+					end
+					@library.build_subdir = (listing.first) if listing.length == 1
+				end
+			end
+			true
 		end
 
 		def untar
